backend/requirements.txt
# Web framework
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# Database / ORM
sqlalchemy>=2.0.29
psycopg2-binary>=2.9.9

# Google Cloud / Vertex AI
google-cloud-aiplatform>=1.49.0
google-cloud-storage>=2.14.0

# Supabase client
supabase>=2.4.0

# HTTP / uploads
httpx>=0.27.0
python-multipart>=0.0.9

# Environment variables
python-dotenv>=1.0.1

# Data validation
pydantic>=2.6.0
email-validator>=2.1.0
